Stop running `kttxt` validation on every PR to `main`
This commit stops running the code snippet validation on any PR to the `main` branch, as the step is very slow (in most of the cases I have faced over 1h).

Instead, the snippets would be executed ONLY when they're changed, and by that, the organization can save expensive runner calculation minutes that are currently used for no good.

Moreover, any PR that does not touch the snippets could be merged to the `main` branch much faster as no validation of the snippets would be required.